Add a task to clean up workflow's closed projects
If something goes wrong during the run of a workflow and the error is
not correctly caught (yet), the workflow_run can stay in 'running'
status forever and the temporary project created by the workflow is not
removed as expected.

Apart from fixing the errors, we need to ensure we clean up those
projects. This rake task is introduced for that purpose.

As the artifacts were never created, we rely on the workflow_run's
payload information to get part of the project name (repository name and
the PR number).