Make `package_find_options` an argument for triggerable concern
Since we are going to use the concern not only for tokens
but also for workflow steps from now on, we need a way to
pass the `package_find_options` from the steps.
Until now it worked out without doing so, but since we are going
to have more than one step that uses the concern we cannot simply
define it in the workflow token anymore.