[backend] fix broken additional metadata with createrepo_c 0.20
createrepo_c changed the default to keep always old metadata.

At the same time modifyrepo_c got changed to fail when the referenced meta data file
got already removed. This leads to the situation that old meta data stays forever.

Not removing the files is not an option as they would stay forever when
actually should not get generated anymore.

Therefore we require the new version of createrepo_c now to be able to switch
back to the old behaviour.