Use ConfirmationDialogComponent in watchlist
To delete an item from the watchlist, we rely on a confirmation dialog
that is now rendered as a component.

As we render the dialog from an external HTML node (to make it cover the
whole page content), we need to pass the infomation about the clicked
item through data attributes.