Prevent code 500 on packages received through an scmsync
We currently receive exceptions in errbit due to packages
that are received through projects that use scmsync.
Those packages are not known by the OBS frontend and set
to `nil`. This crashes in the package#show view, since the code
tries to access package properties on the the nil object.
Therefore we should throw an exception instead of setting
the package to nil.

Fixes #12552