Avoid code 500 on packages received through an scmsync for package show
We currently receive exceptions in errbit due to packages
that are received through projects that use scmsync.
Those packages are not known by the OBS frontend and set
to `nil`. This crashes in the package#show view, since the code
tries to access package properties on the the nil object.

Fixes #12552