[api] protect us against broken requests due to multiple entries
The active record validation is not protecting us here on creation
time, so introducing uniq_keys for validation.

Also we can conflict correctly between different action types this way.
Only action types which do not conflict with others for the same target
have therefore a prefix in the conlflict key.