Toggle visibility of the token secret
In order to effectively share a token, the user that gets the shared
token needs to know the token's secret. As of now, no one but the token
creator knows the token's secret.

This change fixes that, allowing a user with the proper permissions to
show and hide the secret from the token's page.