Set JS code in DeleteConfirmationDialogComponent
When we need to set the content of the dialog dynamically, we pass data
attributes with the values and call the JS function that sets them.

Instead of calling that function from every view, we include the JS code
in DeleteConfirmationDialogComponent, which collects all the modals to
confirm deletion we find in the DOM and set the values coming from the
corresponding data attributes if any.