Introduce validation of EmbargoDate attribute
This means it will be no longer possible to introduce an invalid
EmbargoDate attribute.

Anyway, the step of checking for an embargo date when releasing packages
still handles the case of invalid embargo dates, which were stored
previously in the database. This PR doesn't contain any data migration
to modify embargo dates according to the new validation rules.

As a consequence of introducing this validation in the `Attrib` model,
some specs were modified to make use of invalid `Attrib` objects.
Regarding maintenance minitests, as the case of invalid EmbargoDate
attributes is already tested in specs, the minitests were adapted to
handle the case of valid EmbargoDate attributes.