Test rack/rake gem versions
The system and bundle versions of rake and rack need to match up.

The rake and rack gems are dependencies of passenger that it will load, from the
system, during booting itself. If our bundle versions deviate, then booting our app
will try to load those gems again, which will fail. Which makes for an unbootable
app.