Adapt specs for save_scm_report_* methods
The specs relied on the value returned by the methods
`save_scm_report_failure` and `save_scm_report_success`,
which was the just-created instance of `ScmStatusReport`.

As we are not using the returned value on any other part of the code and
we want to send metrics as the last instruction of the methods, we adapt
the specs accordingly.