Allow users to manage their own home project and its packages
This is needed since users sometimes remove themselves from the
maintainers of their own home project. It should have never been
possible in the first place for users to do this, but here we are. It
was quite confusing to users since they could still see the `Create
Package` or `Add User` actions, but they were never authorized to do
those actions.  They should have been able to do them though. This is
now addressed.

For now, we can help users by applying this fix. This code from the
custom authorization system is called in Pundit, so both authorization
systems (custom and Pundit) are covered.

Another important change here is that nothing can prevent a user from
managing their home project or a package in their home project. As an
example, it was previously possible to prevent users from accessing the
source of their home project by setting `sourceaccess` to `disable` in
the local permissions.

Fixes #10067