[backend] support sending of a large number of obscpio files
We used to pre-open obscpio files if we are building a cpio list.
This has the disadvantage that we run into the filesystem limit
if we need to send a large number of obscpio files.

This commit adds support for passing a function reference as
file handle to BSCpio and BSTar. We use this to lazily open
obscpio files from the source repository.