Simplify require_request method in Webui::RequestController
`require_request` is used in a before_filter callback for
`changerequest`, `show` and `request_action` actions. The request number
has to be passed in the `show` and `request_action` routes, so it's
impossible to not have `number` in the `params`. As for `changerequest`,
the request number is also passed in the `params` via a hidden field
when the form is submitted. So unless a user changes the HTML, it's
always going to be there. And if this happens, there's nothing we can
do.

This explains why relying on required_parameters and all this custom
code is not needed. We can instead do it the Rails way and rely on
`find_by!`, which raises `ActiveRecord::RecordNotFound` if the request
isn't found, then the controller responds with 404.