Remove calls to start the test backend
This call is only needed when running the minitests.

Once the backend test is up, it keeps running until minitests are
finished.

Add a call to start the test backend in the first minitest from the
group of tests run in the `test:api:group1` group. This makes it
possible to run this whole group of tests with a test backend.