Prevent ActiveRecord reflection error in Cloud::* models
The error:

```
ArgumentError:
Rails couldn't find a valid model for User association. Please provide the :class_name option on the association declaration. If :class_name is already provided, make sure it's an ActiveRecord::Base subclass.
```

Specifying the User model in those Cloud::* models prevent clashes with
the Cloud::User module, which is why ActiveRecord was producing this
error.