Display review's project/package name as string
Sometimes the package associated with the review is deleted, i.e. after
accepting a deletion request. So we stop retrieving those packages and
simply display their names as a string

We do something similar with the deleted projects. Before we didn't
display the associated review at all.

In those cases, as we can't retrieve the associated users, we display a
package or a project icon as avatar.