Stop handling linkinfo in sourcediff components
The partial `_revision_diff_detail` is called from the package's
controller action `rdiff` (alias `package_rdiff`) or from the request
view.

In the first situation, we are setting @linkinfo on the package
controller and then it is passed to the partial.

In the second situation, we never set @linkinfo on the request controller.
So there is no need to pass @linkinfo to the component in order to always
end up passing nil to the partial.