Pass active action to RequestDecisionComponent
The only reason why we pass the whole collection of actions to the
component is to count them inside the `single_action_request` method
and to get the active action.

From now on, we stop passing @actions and we just pass the active action
and we count the number of actions using a Rails query.

Co-authored-by: Rubhan Azeem <rubhan.azeem@suse.com>

temp