Remove filters usage from ReportToScmJob
At the moment of reporting back to the SCM, we used to take only two of
the workflow filters into account: `architectures` and `repositories`.
With them we only reported information about builds related to the
specified architectures and repositories.

We no longer need that functionality, so we stop dealing with those
filters in ReportToScmJob. We remove the related code and specs.