Check pull/merge request action against allowed actions
We already have a constant for supported pull/merge request
actions (allowlist). There is no need to create a blocklist
for ignored actions. We can simply make sure the action is not
included in the allowlist. This makes it easier for future
implementations (Gitea), since we don't have to check all possible
actions.