Stop passing around workflow filters hash
The hash of workflow filters was passed to Step#call and beyond to end
up storing it in the subscriptions payload. That was the way to make the
hash (`architectures` and `repositories` filters) be available for the
ReportToScmJob later on.

Now that the ReportToScmJob is no longer dealing with those filters, we
stop passing them. As the other filters `branches` and `events` are
checked only before calling Step#call, in the Workflow#call model.