Create rake tasks to migrate the old watchlist
Add two rake tasks meant to be used by OBS admins, who want to make
easier for OBS intances' users the transition from the project watchlist
to the new watchlist.

The first task 'Create a list of emails which should be used to warn
users' will retrieve the emails from users which, at the same time they
have projects in the new watchlist, also have some projects in the old
watch list which aren't in the new watchlist.

The second task 'Copy projects from the old watchlist to the new one'
copies the projects that the user has in the old wachtlist if the user
doesn't have any project in the new watchlist.

This last task is irreversible. If you want to warn users by email,
run the first task to retrieve the list of these emails and send the
email. Afterwards run the last task.

Co-authored-by: Daniel Donisa <daniel.donisa@suse.com>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>
Co-authored-by: Henne Vogelsang <hvogel@opensuse.org>