Move Package.download_url_for_binary back to controller
We need to account for packages that come through project links. The only
reasonable place you can know about this is the controller. Where the Project
and the Package that comes through a project link are instantiated.

An instantiated `Package` doesn't know which `Project` it was instantiated with
in `Package.get_by_project_and_name`...

Fixes #13147