Creates a link to each section of a page ...
... with nab-tabs Bootstrap elements through fragment identifiers. Only
for the request page of the beta program.

Use a fragment identifier in the url (`#` and a section identifier),
instead of using the `tab_name` parameter. Don't use any fragment
identifier for the first tab option.

Adapt the action selector accordingly, adjusting the `href` attribute of
the action links every time the actions dropdown is selected.