[api] Avoid incomplete releases via token
It failed with an absurd error in the middle of a release when another
releasetarget lacked the permissions.

The `release_package` method of the `MaintenanceHelper` mixin checks
for permission on the given release target/project. But to avoid
incomplete releases, we need to check all release targets of the
collection upfront, before we start to release them one by one.