Fix scmsync tag creation for packages with query/fragment in the url
The SCM workflow creates bogus scmsync URLs when the branched package contains
any query or fragment in its url, because neither the query nor the fragment are
ever removed/modified. Only the new commit is appended as a fragment which can
lead to duplication of the query & fragment.
To fix this, we use the addressable module to override the fragment and set the
`subdir` query via the URI parser instead of doing this ourselves.

This fixes https://github.com/openSUSE/open-build-service/issues/13290