Move code closer to where it will be used
We don't really need to calculate those variables in the controller
just to push them down over several partials.

Bringing that variable creation closer to where it will be used will
reduce the likelihood of interfering with future changes.

As a side note, the creation of those variables in the _conversation
partial hints at the creation of a View Component to hold the
conversation view.