Rename rpm-lint-result to rpm-lint
There is content of the "RPM Lint" tab that is static, only rendered
when the page is loaded. And there is some other content that is
rendered via AJAX.

To prevent confusion on what is retrieved with AJAX (the result), rename
the references to the content of the tab to rpm-lint and leave
rpm-lint-result for the content related to each result.