[backend] support multiple sandbox constraints
We may have a definition in prjconf and in the package. Usually they
conflict, but if one of them is the "secure" one and may also be
good with a package level constraint for sandbox.

This is a change which can be incompatible in rare cases, eg.
on project level "secure" was required and on package level
"kvm" is enforced. This used to get assigned to "xen" workers,
which is now not accepted anymore.