Handle old and new receiver_roles when updating the event subscription
In order to support both the new and old naming for the receiver_role
`watcher` and `project_watcher` we need to update the old existing
records by just switching the `enabled` field while keeping the old
naming scheme for the receiver_role `watcher`. For non existing
subscriptions we create a new record with the new naming scheme
`project_watcher`.

Co-authored-by: Lukas Krause <lkrause@suse.de>