Render checkboxes for existing event subscription receiver_role name `watcher` correctly
In order to migrate the `receiver_role` names from `watcher` to
`project_watcher` we have to support both implementations
at the same time. In order to do so we have to keep rendering the
checkboxes correctly for the old role names. We going to update
the existing records for with the old role names. For all non
existing ones, we going to use the new naming.