Set correct commit sha for github on tag push for target_branch
The json payload key `:after` in the webhook of a push event
from GitHub contains the sha of the tag itself. Instead we need to use the
commit sha the tag is refering to. This is part of `:head_commit =>
:id`. The `:target_branch` payload field is used in the yaml_downloader
to generate the url in order to pull the workflow configuration yaml
file.

Fixes #13206