Update Puma to fix chunked upload failures
Older verisons of Puma had a bug where chunked uploads would result in
an empty body:

https://github.com/puma/puma/issues/1839

This breaks OBS file uploads that rely chunked transfers, which is
generally the standard mechanism used by http clients for uploading
large files. A fix was backported to 4.3.7:

https://github.com/puma/puma/pull/2496

so this PR updates Puma to the latest 4.3.x release to include this
fix.

Likely fixes #9329.

(Note that, if OBS is behind a reverse proxy, this issue may not be
observed in the first place, as many reverse proxies will tend to
buffer the uploaded data, thus side-stepping the use of chunked uploads
entirely.)

Signed-off-by: Ryan Gonzalez <ryan.gonzalez@collabora.com>