Set correct commit_sha on tag push for github in scm payload
We use the `:commit_sha` in the branch_request file to
update/overwrite the one used in the `_service` file for the
branch_package and link_package step in order to pull the updated/pushed
sources. This is generated in the steps model with the
`branch_request_content_github` method.
We have to point here to the commit the tag is refering
to, not the sha of the tag itself. Otherwise the service
will fail to pull the sources since there in no correct reference
to the commit.