Use correct commit_sha for gitea on tag push
Similar to github, we have to use the `:head_commit => :id` attribute
from the webhook payload on a tag push in order to point to
commit the tag is refering to, not the sha of the tag itself.
Otherwise we are not able to pull the workflow configuration and the obs_scm service
won't pull the correct sources since the wrong commit_sha is used in the branch_request
file.