Reject set_flags step when deleting projects after a PR was closed
There is no project that should be deleted for the set_flags
step in a workflow, when a PR is closed on the SCM.
There for it should be one of the steps that gets rejected when
we clean up. This will also fix an issue, since the `target_project_name`
method of the `SetFlags` expects to get passed a keyword argument,
which all other steps don't.

Fixes #12637