Correctly check for the current page in request breadcrumbs
Previously, it would only match on URLs like `/request/show/14`, but not
on `/request/show/14/request_action/26`. The second URL example is what
we get whenever selecting an action in the `Select Action` dropdown of
the request#show page with the `request_show_redesign` feature toggle.

Now, breadcrumbs are displayed for both URLs.