Reject set_flags step when restoring projects after a PR was reopend
There is no project that should be restored for the set_flags step in a workflow,
when a PR is reopened on the SCM. There for it should be one of the steps that gets
rejected when restoring. This will also fix an issue, since the target_project_name method
of the SetFlags is private and expects to get passed a keyword argument, which all
other steps don't.

Fixes #13338