Extract the correct reference to the YAML file
The workflows YAML file should be downloaded from the target branch or
commit, not the source one.

We want to rely on the head commit sha of the target branch.
To do so, we add a new key in the SCMWebhook payload:
workflow_configuration_ref.

In case of a push event that deletes the ref, we need to take the commit
SHA from the "after" key comming from the original payload.