Use Marcel::Magic instead of hardcoded binary extensions
Using this gem because it's self contained, already pulled in by rails
and allows us to append more types than are included in its default
tables. There is MimeMagic which this gem is based on, which uses a more
complete Freedesktop database for mime info, so we likely wouldn't need
to add the initializer code, but it would require adding a gem and its
system dependencies, for very little benefit. If however there is ever
a need to switch over to it, it's very similar in structure, so it
would be fairly easy to do.