[backend] avoid service marks without server side services
Checking if any service is actually configured to be
executed on the server side. Avoid creating a service
mark and service run in that case.

A small, but wanted side effect is that service commands are
failing now, when a service is defined, but not active (like
with no service defined at all).