Introduce limiting results for /search/... endpoints
Add the `limit_for_search_results` configuration variable, to limit the
results returned by calls to API endpoints starting by `/search/...` and
not ending in `.../id`.

Not defining the configuration variable results in not changing the
behaviour of the endpoints. That is, the results are not limited.

If the number of results surpases the value of the configuration
variable, an 403 error is thrown, with a text advising of the possible
alternatives to consider:

- retrieve only the ids by using an '/search/.../id' API endpoint, or
- reduce the number of matches of your search:
  - paginating your results, through the 'limit' and 'offset' parameters, or
  - adjusting your `match` expression.