[backend] Kiwi product: move seen_fn test after setting seen_binary
This further reduces the number of rpms we need to download. The
old way of doing the seen_fn check first led to seen_binary not
being set if a noach package was taken from a better architecture.
This in turn led to rpms being taken from older packages.

I.e. we took the noarch package with the highest version from x86_64,
an older version from s390x, the next version from ppc64le and so on.