Remove enforcing `vcr: true` for feature specs
VCR was enforced for feature specs. This happened trying to make VCR the
default for feature specs.

Instead of fixing the code and making VCR the default for feature specs,
remove enforcing of VCR for feature specs. This way, the `vcr:` flag is
used like in the rest of the tests, and for those tests which don't need
VCR there is no need to use `vcr: false`. We simply don't define any
`vcr:` flag.