Split github action workflows into two separate workflows
In order to allow the creation of comments on PR's that are
coming from forks, we would need to use the `:pull_request_target`
event over the `:pull_request` event. Since the `:pull_request_target`
event comes with more permissions on the target repo, its not recommended to
checkout the code coming from an untrusted fork.
Therefore the gh action workflow is now splited into two separate
workflows.
One workflow keeps working on the `:pull_request` event and checksout
the code to perform the diff operation in order to receive gem
name, version etc. Afterwards the informations are saved as artifacts.
The second workflow listens for the `workflow_run` event and gets
triggered after the first one finished successfully, then downloads
the artifacts and finally creates the comment (the `:workflow_run`
event has permission to comment on the PR).