Move longer tests in a row to the end
RuboCop linter tests take longer than other linters.

Move them to the last position of the sequence of linters. This way,
linters fail sooner because the linters that take less time are run
first, and therefore, CI steps fail sooner on average when we encounter
a linter offense and we use less test cycles.