Hand over PR number between gh action workflows through artifacts
The PR number is not available in the payload of the `workflow_run`
event, when the PR is coming from a fork. To workaround that problem
I write the PR number to the artifacts in the first workflow that gets
triggered from the `pull_request` event (where the pr number is
available) and fetch it in the following one from there.