Create ruby script to resolve references in swagger yaml
In order to speed up the rendering of the OpenAPI documentation
in Swagger, we need to resolve all the references to the different
yaml files and create on big yaml file. This improves the loading
time a lot since we only have on query instead of multiple.

Local tests without caching:

Before: 3,69sec -> 268 requests in total
After: 172ms -> 8 requests in total

We will use the script during the build process of the obs-server
package to replace the Swagger YAML that we use during development
with the YAML that inlcudes all resolved references to the other
YAML files.

Usage:
```
./resolve_swagger_yaml.rb -i ../src/api/public/apidocs-new/OBS-v2.10.50.yaml -o ../src/api/public/apidocs-new/OBS-v2.10.50_new.yaml
```
OR
```
./resolve_swagger_yaml.rb -i ../src/api/public/apidocs-new/OBS-v2.10.50.yaml -o ../src/api/public/apidocs-new/OBS-v2.10.50.yaml -f
```
to force overwrite of an existing file