Stop adding branch_request files for linked packages
The backend changed how it decides if service should run for a package.
Services in modes that would not run server side are filtered out.

We used the project wide service in localonly to "trick" the backend into
running services for linked packages in the link_package step.

As this stopped working a while ago, services are not running anymore for links.
Which makes them kind of useless for PR events where it's supposed to run the
services to pick up changes from the PR.

Pepple will have to use branch_package instead that will create a local _service
file.