Move cache invalidation into background job
There are manual caches in `User.tasks`/`Group.tasks` and also fragment caches
in views that use `cache_key_with_version` which we invalidate by touching. For
a `BsRequest` with many `BsRequestAction` or a `BsRequestAction` for a
`Project` with many `User` this can take quite a while.

This shouldn't happen in a callback as saving a `BsRequest` often runs enclosed
in (manual) transactions...