Make the error ingoring simpler
We don't need to check for the presence of `<summary>` because in line 79 we check
that errors are of Backend type, so any message not having `<summary>` is not a
Backend error so it won't enter the matching function.

Once we don't have the `<summary>` tag, we don't need to check with regexp because
we can just check for part of the message.