[webui] Default expand=0 when revision is set
For packages that are branches, the previous expand=1 default
meant that the (newer) linked package was shown
instead of the old revision that was requested.
This violated the principle of least surprise.

For packages that are not branches, behaviour remains the same.

For packages that were created with osc linkpac, the view will change.

Fixes #12674

Co-Authored-By: Bernhard M. Wiedemann <bwiedemann@suse.de>