Tell people about Pundit errors in WorkflowRun
We started to use Pundit a lot in the various `Workflow::Step`.
Authorization failures that happen inside the `Workflow` should
be reflected in the `WorkflowRun` and not in the request cycle.

Also: As the `TriggerWorkflowController` takes care about APIErrors
and Pundit now, we don't need to handle those exceptions in
`Workflow.destroy_target_projects`.