Fix button visibility condition
One of the conditions is enough to hide the show_all button

Scenario fixed: when pagination pages are just 1 and the user is logged
in, the condition was returning `false` because of the `&&`,
so the show_all button was visible. But this is wrong because
there is no pagination and all the revisions are already visible in the first page.