Fix reusing EventSubscription
Including the `WorkflowRun` in the `find` part we will never find a
previous `EventSubscription` because we are always in a new `WorkflowRun`
when we are using this service.

So we always `create` a new `EventSubscription` and we end up having one subscription
per SCM event. This makes us report for *each* commit and in case of PR synchronization
even twice for the same commit...