Test for possible version mismatches of base64 (stdlib) gem
base64 is part of the ruby stdlib https://stdgems.org/base64/.
Rubocop adds it now as a gem dependency since it gets removed
from the stdlib with ruby 3.4 rubocop/rubocop#12094.
The introduced version of base64 matches the one of our
currently used ruby version (3.1.4). We have to make sure
that the version that now is included to our bundle does not
mismatch in the future due to possible introduced updates.