[api] implement package fork command
This can be used to create a test build setup for a scmsync package
based on an existing project.

The base package may or may not exist. May or may not
use scmsync.

This is mainly implemented inside of BranchPackage class
intentionaly, even when this is no source branch technically.
But many api parts like repository setup can be shared this way.

Co-Author: Henne Vogelsang <hvogel@opensuse.org>