Align behaviour of user `delete!` and `mark_as_spammer!`
Those methods are both setting the `state` of a user to `deleted`
but have slight differences. The `mark_as_spammer!` also deletes
the comments associated with the user, while the `delete!` method
clears the users `realname` and `email`. To keep it simple we
shoud align those and maintain only one.