Fix permission checks and crashes for any situation where the
package is not inside of the local project. yes, this includes also
project scmsync situations, but also project links to either local or
remote.

multibuild flavor packages are NOT the reason when @package was
a string either, so dropping the obsolete code doing a second
lookup in local project only.

Instead keep global @package_name to be able to handle situations where
a remote package, scmsync managed project or multibuild flavor situation
needs to be handled.

Adding missing permission check for the project access also, since the
package may not be part of the project eg when building via project
links.