Improve minimum button size of requests link in tables
The current request listing requires to hit *exactly* the 15x18 pixel
region to the icon to get into the detailed listing. nothing else
in the whole line links to details. there is no sensible keyboard
navigation either.

This improves it to roughly 30x36 which is still far away from the
minimum recommended sizes for user interaction elements (44x44) but
slightly less insane.