Move Report component call outside the comment list
Project, package and request pages use the comment list. The project and
package pages already call the Report component. The request view didn't.

Call the Report component from the `request/_request_comments partial`,
so we can get rid of the call of the Report component in the
`comment/_show` partial, wich is redundant for projects and packages.