Fix javascript and remove render condition from report modal
We cannot authorize inside the render condition in this case.
We render the modal on the package, project and request views,
then set the attributes on the form inside the model via
javascript. When for example we cannot report on the package itself
but on the comment that is on the package view, the component
is not rendered. We already have authorization on the controller
code and the report links. There is no need to do this on the modal.