Remove unreachable code in build/file controller
When there is a query string in the url, `request.query_string` returns
the query string, which is considered `true` in the `if` conditional.

When there isn't any query string in the url, `request.query_string`
returns the empty string `''`, which is also considered `true` in the
`if` conditional.

Therefore, the code after the guard clause inside the `regexp` method is
never reached, and also calls to the `regexp` method always return `nil`
making the check `if regexp` return `false` and always enter the `else`
section.