Drop bogus code blocking access to build log
-        # NOTE: @package is a String for multibuild packages
-        @package = Package.find_by_project_and_name(@project.name, Package.striping_multibuild_suffix(@package_name)) if @package.is_a?(String)

The comment is wrong, a multibuild flavor would not lead to a string. A
remote or scmsync package would cause this.

also the find would not find local packages via local project links.
But in that case it would actually open a security issue, since it
would not perform the right permission check.

This did only not cause this issue due to crashing right after.