Fix mocking a request to the backend in trigger..
... controller specs

We should have mocked a POST request, not a GET request.

Mock the call to the Backend::Api library, instead of mocking an
explicit HTTP request to the backend. Now we don't need to know what
HTTP action we should use (POST or GET), and we can get rid of lots of
cassettes.