Fix `create` action of Webui::GroupsController
Fix the `replace_members` Group model method and its specs: prevent
removal of previous groups users when for, at least one user of the new
members to replace, there are validation failures or the user is not
found.

Create a transaction when saving group details and group users in the
`create` action of the Webui::GroupsController. This helps to guarantee
not creating a group when there are group details validation failures,
or new users validation failures or not found new users.

Adapt specs of `create` action of the Webui::GroupsController.

Found while trying to fix a RSpec/InstanceVariable RuboCop offense, also
removed from the RuboCop TODO file.