Adjust position of Reply and Report links
Make Report and Reply links to stay in the same position and make the
collapsible form open below them.

Also move links to the right side to minimize the visual noise on the
left, where there are many other elements like lines, icons and
avatars. The right side is the place where the reader's eyes usually
end up when they finish reading.