Add List-Id for better filtering with GMail
GMail is not able to filter emails by arbitrary headers, but only by
a few selected ones. the easiest one to use is "List-ID", which is also
set by e.g. GitHub for notifications in the notifications.

Setting such a header allowes in combination with optionally parts of
the subject very powerful and reliable filtering of notifications when
a user would be using GMail or Google Workspace.