Introduces triggering Project operations
There are two types of trigger routes:

On /trigger you can trigger an type of `Token`
On /trigger/[rebuild, release, runservice, workflow] you can trigger only this type of `Token`

If a `Token` has a `Package` the operation is executed on this `Package`.

If a `Token` does not have a `Package` you can set one via the project/package
parameters and the operation is executed on this `Package`.

If a `Token::Rebuild` or a `Token::Release` has no `Package` you can set a
`Project` via the project parameter and the operation is executed on this
`Project`.

Co-authored-by: Adrian Schröter <adrian@suse.de>
Co-authored-by: Eduardo Navarro <enavarro@suse.com>