Instantly display the 'You reported the X' message
A page refresh isn't needed anymore for the message to appear.

For comments, the implementation isn't really robust, but considering
how the comments are rendered, this is what we can do without
refactoring a lot of code. In the best scenario, rendering a comment
would be as simple as rendering a single view component. It would then
be unneccessary to fiddle much with JavaScript, re-rendering the comment
view component would be the solution.

As for projects and packages, it's not possible to re-render the
side_links partials since they depend upon so many objects and we don't
have those in webui/reports_controller.