[backend] BSWatcher: add serialize_waiting() and serlialize_advance() functions
The serialize_waiting() function returns all events waiting for a
serialized resource.

serlialize_advance() can be used to move an event to the front
of the waiting queue.

Both functions will be used in the next commit to make requests
for the same resource only do one request. The idea is to
check if waiting requests ask for the same thing and then donate
the result to them so that they have not to do the same rpcs.
This is more lightweight than a cache and also does not need
code for cache expiry.