Remove validation of token type
This validation made `osc` throw an exception when using `osc trigger`
without the `--operation` parameter over a release token.

The reason is that `osc` calls `POST /trigger/runservice` by default
when the operation is not passed. Even if the `Token#type` is a
different one.

We want to introduce this validation, but not now. We will wait until a
new version of `osc` is released, with the support of the `POST
/trigger` API endpoint.