Authorize comment `#create` api endpoint
Currently there is not authorization going on, in some cases
the `CommentLockingValidator` still prevents comments from being
created if permission is insufficient in certain cases.
But some things are not handled, for example when a user is
blocked for commenting.

This should be handled on the controller level using our pundit
policy.