[api] relax package meta
OBS is not always delivering title and description elements itself
anymore (when they come from pure backend projects). So accept
package meta also without it to avoid that tooling can not
read and write same meta from OBS instances.

Also the order of title and description does not matter to us.

Rest is just indenting fixes.

Adding a test case to ensure that api and backend validation
is able to deal with minimal meta